run_openscap_scan: Initial Commit
Ansible playbook that runs the appropriate OpenSCAP scan using the standard security profile against a host(s).  After the scan is complete the playbook generates an Ansible remediation playbook to address the items that are not in compliance with the security profile.  The scan result along with the remediation playbook are sent out via email.